+++ pdf-text-spring-boot-starter/src/main/java/io/xpdftools/pdftext/autoconfigure/PdfTextToolProperties.java
@@ -18,6 +18,6 @@ public class PdfTextToolProperties {
 
     private Path nativeLibraryPath;
     private Path defaultOutputPath;
-    private Long timeoutSeconds;
+    private Integer timeoutSeconds;
 
 }
